LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/browsers/the-window-object/BarProp.window-expected.txt


FAIL BarBrop objects of a nested Window assert_equals: a BarProp's visible is wrong expected true but got false
FAIL BarProp objects of an auxiliary Window assert_equals: a BarProp's visible is wrong expected true but got false
